sam: standardize samd51 uart settings
